ensure inventory integrity and input validation

+++ backend/internal/dto/report.go
@@ -3,8 +3,8 @@ package dto
 type TurnoverReportItem struct {
 	ProductName  string `json:"product_name"`
 	SKU          string `json:"sku"`
-	StartBalance int    `json:"start_balance"`
-	Income       int    `json:"income"`
-	Expense      int    `json:"expense"`
-	EndBalance   int    `json:"end_balance"`
+	StartBalance int64  `json:"start_balance"`
+	Income       int64  `json:"income"`
+	Expense      int64  `json:"expense"`
+	EndBalance   int64  `json:"end_balance"`
 }
